feat(products): extract ProductsClient and keep page server-rendered
